Bump netdisco to netdisco_wrapper
